ubuntu/code/requirements.txt
kafka-python
psycopg2
faker
requests
sqlalchemy